Sticky RPM!!!
added logic that does the following:

When currentState is SHOOT, the target RPM will be repeatedly recalculated with a valid tagPose data (which already happens) and the target RPM generated by flywheelSpeeds is saved in a variable. If tagPose is not valid, but a previous valid RPM has been saved during the current SHOOT state, it will set the RPM to the stickyTargetRPM. If nothing has been saved, however, the shooter will be set to the defaultRPM as usual.
Whenever the currentState changes to something other than SHOOT, stickyRPM is reset.

This should help with the issue we seem to be having occasionally when shooting from the far launch zone where the robot intermittently detects the April Tag, resulting in a fluctuating target RPM and inconsistent shooting.