Fixed the "multi-shoot" log glitch
I basically just fixed the weird thing in logs that @FreehandBlock51 and I saw where the robot would say it shot 3 times when it only did 2. I think the reason it was happening earlier was because the tolerance was too small, so the rpm would occasionally oscillate outside tolerance without anything actually happening. Increased tolerance from 100 to 150, and haven't had that issue since (in several runs).
The main issue with auto now is the timing. Potentially we could combine some of the multi-step intake paths into single movements to reduce the amount of accelerating and decelerating that the robot has to do, which might shave off the necessary 2-3 seconds (from what I have seen we are getting 11 artifacts pretty consistently, so we need time for the 12th and leave).
please excuse the mammoth commit description for a 1-line change, I just want to document what I noticed with the robot